# openapi-swagger.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Order APP API",
    "description": "API for managing orders in the Order APP",
    "contact": {
      "name": "Luan Ricardo Gazin",
      "url": "https://www.linkedin.com/in/luangazin",
      "email": "luan_gazin@yahoo.com.br"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Server"
    }
  ],
  "tags": [
    {
      "name": "Order Management",
      "description": "Operations related to order management"
    }
  ],
  "paths": {
    "/v1/orders": {
      "get": {
        "tags": [
          "Order Management"
        ],
        "summary": "Search orders",
        "description": "Searches for orders with optional filters",
        "operationId": "searchOrders",
        "parameters": [
          {
            "name": "order-id",
            "in": "query",
            "description": "Order ID",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "partner-id",
            "in": "query",
            "description": "Partner ID",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Order Status",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "APPROVED",
                "PROCESSING",
                "SHIPPED",
                "DELIVERED",
                "CANCELED"
              ]
            }
          },
          {
            "name": "start-date",
            "in": "query",
            "description": "Start date",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "end-date",
            "in": "query",
            "description": "End date",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Page size",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "sort-by",
            "in": "query",
            "description": "Sort field",
            "required": false,
            "schema": {
              "type": "string",
              "default": "createdAt"
            }
          },
          {
            "name": "sort-direction",
            "in": "query",
            "description": "Sort direction",
            "required": false,
            "schema": {
              "type": "string",
              "default": "DESC",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order list found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListOrderResponseDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Order Management"
        ],
        "summary": "Create new Order",
        "description": "Creates a new order for a partner",
        "operationId": "createOrder",
        "parameters": [
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Idempotency key to safely retry requests",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderPostRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Order created successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOrderResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad request, invalid input",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOrderResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Not found, partner not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOrderResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOrderResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/v1/orders/{id}/status": {
      "patch": {
        "tags": [
          "Order Management"
        ],
        "summary": "Patches order\u0027s status",
        "description": "Patches an order\u0027s status",
        "operationId": "updateOrderStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Order id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "New status",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "APPROVED",
                "PROCESSING",
                "SHIPPED",
                "DELIVERED",
                "CANCELED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status updated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOrderResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad request, invalid status",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOrderResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOrderResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/v1/orders/{id}": {
      "get": {
        "tags": [
          "Order Management"
        ],
        "summary": "Find order by id",
        "description": "Returns an order by its ID",
        "operationId": "getOrderById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Order id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOrderResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOrderResponseDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Order Management"
        ],
        "summary": "Cancel order",
        "description": "Cancels a specific order",
        "operationId": "cancelOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do pedido",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Order canceled successfully"
          },
          "400": {
            "description": "Order cannot be canceled"
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "OrderItemDTO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number"
          }
        }
      },
      "OrderPostRequestDTO": {
        "type": "object",
        "properties": {
          "partner-id": {
            "type": "string",
            "format": "uuid"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemDTO"
            }
          }
        }
      },
      "ApiResponseOrderResponseDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "SUCCESS",
              "ERROR"
            ]
          },
          "error-code": {
            "type": "string"
          },
          "error-message": {
            "type": "string"
          },
          "body": {
            "$ref": "#/components/schemas/OrderResponseDTO"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
      "OrderResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "partner": {
            "$ref": "#/components/schemas/PartnerDTO"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemDTO"
            }
          },
          "totalValue": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "PROCESSING",
              "SHIPPED",
              "DELIVERED",
              "CANCELED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "total-items": {
            "type": "integer",
            "format": "int64"
          },
          "total-pages": {
            "type": "integer",
            "format": "int32"
          },
          "current-page": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PartnerDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "ApiResponseListOrderResponseDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "SUCCESS",
              "ERROR"
            ]
          },
          "error-code": {
            "type": "string"
          },
          "error-message": {
            "type": "string"
          },
          "body": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderResponseDTO"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      }
    }
  }
}